+++ packages/orq-rc/RELEASES.md
@@ -1319,3 +1319,13 @@ Based on:
 - [python v3.6.0-rc.17] packages/orq-rc
 ### Releases
 - [PyPI v3.6.0-rc.17] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.17 - packages/orq-rc
+
+## 2025-05-20 13:22:11
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.548.1 (2.605.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.6.0-rc.18] packages/orq-rc
+### Releases
+- [PyPI v3.6.0-rc.18] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.18 - packages/orq-rc